Move cmd/aws-actuator/utils under cmd/aws-actuator
The package has no use other than provider CreateActuator for
cmd/aws-actuator which can be moved directly to the package.
