+++ cmd/aws-actuator/utils.go
@@ -1,0 +1,151 @@
+package main
+
+import (
+	"bytes"
+	"fmt"
+	"html/template"
+	"io/ioutil"
+	"os/exec"
+
+	yaml "gopkg.in/yaml.v2"
+
+	"github.com/golang/glog"
+
+	apiv1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/wait"
+	"k8s.io/client-go/tools/record"
+	"sigs.k8s.io/controller-runtime/pkg/client/fake"
+
+	"github.com/openshift/cluster-api-actuator-pkg/pkg/e2e/framework"
+	machineactuator "sigs.k8s.io/cluster-api-provider-aws/pkg/actuators/machine"
+	"sigs.k8s.io/cluster-api-provider-aws/pkg/apis/awsproviderconfig/v1alpha1"
+	awsclient "sigs.k8s.io/cluster-api-provider-aws/pkg/client"
+	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
+)
+
+type manifestParams struct {
+	ClusterID string
+}
+
+func readMachineManifest(manifestParams *manifestParams, manifestLoc string) (*clusterv1.Machine, error) {
+	machine := &clusterv1.Machine{}
+	manifestBytes, err := ioutil.ReadFile(manifestLoc)
+	if err != nil {
+		return nil, fmt.Errorf("unable to read %v: %v", manifestLoc, err)
+	}
+
+	t, err := template.New("machineuserdata").Parse(string(manifestBytes))
+	if err != nil {
+		return nil, err
+	}
+	var buf bytes.Buffer
+	err = t.Execute(&buf, *manifestParams)
+	if err != nil {
+		return nil, err
+	}
+
+	if err = yaml.Unmarshal(buf.Bytes(), &machine); err != nil {
+		return nil, fmt.Errorf("unable to unmarshal %v: %v", manifestLoc, err)
+	}
+
+	return machine, nil
+}
+
+func readClusterResources(manifestParams *manifestParams, clusterLoc, machineLoc, awsCredentialSecretLoc, userDataLoc string) (*clusterv1.Cluster, *clusterv1.Machine, *apiv1.Secret, *apiv1.Secret, error) {
+	var err error
+	machine, err := readMachineManifest(manifestParams, machineLoc)
+	if err != nil {
+		return nil, nil, nil, nil, err
+	}
+
+	cluster := &clusterv1.Cluster{}
+	{
+		bytes, err := ioutil.ReadFile(clusterLoc)
+		if err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("cluster manifest %q: %v", clusterLoc, err)
+		}
+
+		if err = yaml.Unmarshal(bytes, &cluster); err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("cluster manifest %q: %v", clusterLoc, err)
+		}
+	}
+
+	var awsCredentialsSecret *apiv1.Secret
+	if awsCredentialSecretLoc != "" {
+		awsCredentialsSecret = &apiv1.Secret{}
+		bytes, err := ioutil.ReadFile(awsCredentialSecretLoc)
+		if err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("aws credentials manifest %q: %v", awsCredentialSecretLoc, err)
+		}
+
+		if err = yaml.Unmarshal(bytes, &awsCredentialsSecret); err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("aws credentials manifest %q: %v", awsCredentialSecretLoc, err)
+		}
+	}
+
+	var userDataSecret *apiv1.Secret
+	if userDataLoc != "" {
+		userDataSecret = &apiv1.Secret{}
+		bytes, err := ioutil.ReadFile(userDataLoc)
+		if err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("user data manifest %q: %v", userDataLoc, err)
+		}
+
+		if err = yaml.Unmarshal(bytes, &userDataSecret); err != nil {
+			return nil, nil, nil, nil, fmt.Errorf("user data manifest %q: %v", userDataLoc, err)
+		}
+	}
+
+	return cluster, machine, awsCredentialsSecret, userDataSecret, nil
+}
+
+func createSecretAndWait(f *framework.Framework, secret *apiv1.Secret) error {
+	_, err := f.KubeClient.CoreV1().Secrets(secret.Namespace).Create(secret)
+	if err != nil {
+		return err
+	}
+
+	err = wait.Poll(framework.PollInterval, framework.PoolTimeout, func() (bool, error) {
+		_, err := f.KubeClient.CoreV1().Secrets(secret.Namespace).Get(secret.Name, metav1.GetOptions{})
+		return err == nil, nil
+	})
+	return err
+}
+
+// CreateActuator creates actuator with fake clientsets
+func createActuator(machine *clusterv1.Machine, awsCredentials *apiv1.Secret, userData *apiv1.Secret) *machineactuator.Actuator {
+	objList := []runtime.Object{machine}
+	if awsCredentials != nil {
+		objList = append(objList, awsCredentials)
+	}
+	if userData != nil {
+		objList = append(objList, userData)
+	}
+	fakeClient := fake.NewFakeClient(objList...)
+
+	codec, err := v1alpha1.NewCodec()
+	if err != nil {
+		glog.Fatal(err)
+	}
+
+	params := machineactuator.ActuatorParams{
+		Client:           fakeClient,
+		AwsClientBuilder: awsclient.NewClient,
+		Codec:            codec,
+		// use empty recorder dropping any event recorded
+		EventRecorder: &record.FakeRecorder{},
+	}
+
+	actuator, err := machineactuator.NewActuator(params)
+	if err != nil {
+		glog.Error(err)
+	}
+	return actuator
+}
+
+func cmdRun(binaryPath string, args ...string) ([]byte, error) {
+	cmd := exec.Command(binaryPath, args...)
+	return cmd.CombinedOutput()
+}
